Don't start if it is disabled

+++ src/modules/ShortcutGuide/ShortcutGuideModuleInterface/dllmain.cpp
@@ -125,6 +125,11 @@ public:
     virtual void OnHotkeyEx() override
     {
         Logger::trace("OnHotkeyEx()");
+        if (!_enabled)
+        {
+            return;
+        }
+
         if (IsProcessActive())
         {
             TerminateProcess();
